requirements.txt
gymnasium[mujoco]==1.0.0
submitit
numpy
tqdm
hydra-core
omegaconf
wandb
jax[cuda12]==0.6.0
flax
optax 
distrax
nvidia-cublas-cu12==12.9.0.13##temporary fix